requirements.txt
camunda-external-task-client-python3==4.3.0
SQLAlchemy==1.4.39
python-json-logger==2.0.4
prometheus_client==0.15.0
Flask==2.2.2
dnspython==2.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.1
cx-Oracle==8.3.0
pydantic==1.9.1
